src/gmail_client_impl/pyproject.toml
[project]
name = "gmail-client-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yx p", email = "pengyxyx@126.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.123.0",
    "google-auth>=2.28.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
]

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"
